# `(overflowing|checked)_(add|sub)_offset` implementations