Split out from https://github.com/chronotope/chrono/pull/1048.

The main goal is to add `NaiveDateTime::checked_(add|sub)_offset` methods. To quote from https://github.com/chronotope/chrono/pull/1048:

Instead of panicking in the `Add` or `Sub` implementation of `NaiveDateTime` with `FixedOffset`, we need a way to be informed of out-of-range result.

I added the following methods ~(not public for now)~:
- `NaiveTime::overflowing_add_offset` and `NaiveTime::overflowing_sub_offset`
- `NaiveDateTime::checked_add_offset` and `NaiveDateTime::checked_sub_offset`

The `Add` and `Sub` implementations of `FixedOffset` with `NaiveTime`, `NaiveDateTime` and `DateTime` are reimplemented using of these methods. This fixes a code comment:
> this should be implemented more efficiently, but for the time being, this is generic right now.

The best place to put the `Add` and `Sub` implementations for `DateTime` is in the module of `DateTime`, because there they have access to its private fields. I have moved all implementations to the module of their output type for consistency.

-------


Adding an offset works differently from adding a duration. Adding a duration tries to do something smart—but still rudimentary—with leap seconds. Adding an offset should not touch the leap seconds at all. So the methods that operate on `NaiveTime` should be different.

I extracted the part that operates on the `NaiveDate` that could be shared into an `add_days` methods. Previously `NaiveDate::checked_add_days` would convert the days to a `Duration` in seconds, and later devide them to get the value in days again. This now works in a less roundabout way. ~Might~ This would also help with the const implementation later.