@pitdicker 2023-08-11
@djc this PR is now quite small and the new `NaiveDateTime::checked_add_offset` and `NaiveDateTime::checked_sub_offset` make it easier to write code that avoids panics in other parts of chrono.