@pitdicker 2023-09-12
@djc This is a nice PR for us to push over the finish line.
#1071 to fix a panic in `TimeZone::from_local_datetime` depends on it, #1048 to fix other panics, #1270 for const initialization macros, and also a PR I want to open to make methods on `DateTime` const.