@pitdicker 2023-09-23
Added a small optimization to `NaiveDateTime::checked_(add|sub)_offset`: use `succ_opt` or `pred_opt` instead of `add_days`.
This method is used a lot in the `DateTime::with_*` methods to convert the value to local time and back to UTC.

Before the `add_days` optimization https://github.com/chronotope/chrono/pull/1214:
```
bench_datetime_with     time:   [52.090 ns 52.417 ns 52.773 ns]
Found 7 outliers among 100 measurements (7.00%)
  6 (6.00%) high mild
  1 (1.00%) high severe
```
After the `add_days` optimization:
```
bench_datetime_with     time:   [46.121 ns 46.263 ns 46.423 ns]
                        change: [-11.281% -10.739% -10.207%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 8 outliers among 100 measurements (8.00%)
  4 (4.00%) high mild
  4 (4.00%) high severe
```
Now:
```
bench_datetime_with     time:   [34.281 ns 34.365 ns 34.458 ns]
                        change: [-26.354% -25.942% -25.467%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 10 outliers among 100 measurements (10.00%)
  5 (5.00%) high mild
  5 (5.00%) high severe
```

In total we are ca. 35% faster in a good number of methods on `DateTime`.