@djc 2023-09-23
Let's split this in 3 parts?